requirements.txt
pip~=23.2.1
Jinja2~=3.1.3
filelock~=3.13.1
dnspython~=2.6.1
pymongo~=4.6.2
MarkupSafe~=2.1.5
Werkzeug~=3.0.1
click~=8.1.7
blinker~=1.7.0
itsdangerous~=2.1.2
mysql~=0.0.3
mysql-connector-python~=8.3.0
mysqlclient~=2.2.4
virtualenv~=20.25.1
distlib~=0.3.8
platformdirs~=4.2.0
Flask~=3.0.2
python-dotenv~=1.0.1